research/vertica/requirements.txt
vertica_python==1.3.3
fake==0.8
docker==6.1.3